Create new user
Create the user model and its schema.
Add the route with the post method for new users. Add verification to
the route with the 'check' function of express-validator to verify name,
password, and email.
Add the userController, check that the request is not empty with
the 'validationResult' function of express-validator.
Add verification for already registered users.
Add bcrypt to hash the password before saving the new user.